1.0.0.5 增加横幅功能

+++ app/src/main/java/com/ruiduoyi/skyworthtv/model/bean/ProductFragmentBean.java
@@ -8,11 +8,10 @@ import java.util.List;
 
 public class ProductFragmentBean {
 
-
     /**
      * utStatus : true
      * ucMsg : 操作成功！
-     * ucData : {"Table":[{"pqd_xbdm":"K1","pqd_bcdm":"B","pqd_djbh":"700000520661","plm_wldm":"KDB001N0642","itm_pmgg":"KFR-35G/V1BB1A-3","plm_jhsl":1000,"pqd_jhscsl":1000,"rkd_ttl_rksl_v":0,"rkd_day_rksl_v":0,"err_ttl_gzsl_v":0,"err_day_gzsl_v":0,"bzrs":20},{"pqd_xbdm":"K1","pqd_bcdm":"B","pqd_djbh":"700000550940","plm_wldm":"KDA001N1440","itm_pmgg":"HWJZSS50FG01","plm_jhsl":400,"pqd_jhscsl":400,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":500,"err_ttl_gzsl_v":20,"err_day_gzsl_v":10,"bzrs":20},{"pqd_xbdm":"K2","pqd_bcdm":"B","pqd_djbh":"700000520739","plm_wldm":"KDA001W0224","itm_pmgg":"KFR-35W/F2C1A-3","plm_jhsl":1000,"pqd_jhscsl":1000,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":500,"err_ttl_gzsl_v":20,"err_day_gzsl_v":10,"bzrs":20},{"pqd_xbdm":"K2","pqd_bcdm":"B","pqd_djbh":"700000550935","plm_wldm":"KEA001W2120_A4218","itm_pmgg":"SMFH09A-3A2A1NA(O)","plm_jhsl":700,"pqd_jhscsl":700,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":480,"err_ttl_gzsl_v":15,"err_day_gzsl_v":15,"bzrs":20}],"Table1":[{"err_gzdm":"AOIG1","err_gzmc":"AOI工站","err_gzsl":20},{"err_gzdm":"DCT","err_gzmc":"DCT工站","err_gzsl":100},{"err_gzdm":"FCT","err_gzmc":"FCT工站","err_gzsl":0},{"err_gzdm":"ICT","err_gzmc":"ICT工站","err_gzsl":100},{"err_gzdm":"MI","err_gzmc":"MI外观检","err_gzsl":50},{"err_gzdm":"MUTAL","err_gzmc":"自互检","err_gzsl":50}],"Table2":[{"pqd_xbdm":"K1","pqd_jhscsl":1400,"rkd_day_rksl_v":500,"pqd_jhdclv":35.71,"pqd_scxlv":125,"pqd_yjscxlv":6.25},{"pqd_xbdm":"K2","pqd_jhscsl":1700,"rkd_day_rksl_v":980,"pqd_jhdclv":57.65,"pqd_scxlv":245,"pqd_yjscxlv":12.25}]}
+     * ucData : {"Table":[{"pqd_xbdm":"K1","pqd_bcdm":"B","pqd_djbh":"700000520661","plm_wldm":"KDB001N0642","itm_pmgg":"KFR-35G/V1BB1A-3","plm_jhsl":1000,"pqd_jhscsl":1000,"rkd_ttl_rksl_v":0,"rkd_day_rksl_v":0,"err_ttl_gzsl_v":0,"err_day_gzsl_v":0,"bzrs":20},{"pqd_xbdm":"K1","pqd_bcdm":"B","pqd_djbh":"700000550940","plm_wldm":"KDA001N1440","itm_pmgg":"HWJZSS50FG01","plm_jhsl":400,"pqd_jhscsl":400,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":500,"err_ttl_gzsl_v":20,"err_day_gzsl_v":10,"bzrs":20},{"pqd_xbdm":"K2","pqd_bcdm":"B","pqd_djbh":"700000520739","plm_wldm":"KDA001W0224","itm_pmgg":"KFR-35W/F2C1A-3","plm_jhsl":1000,"pqd_jhscsl":1000,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":500,"err_ttl_gzsl_v":20,"err_day_gzsl_v":10,"bzrs":20},{"pqd_xbdm":"K2","pqd_bcdm":"B","pqd_djbh":"700000550935","plm_wldm":"KEA001W2120_A4218","itm_pmgg":"SMFH09A-3A2A1NA(O)","plm_jhsl":700,"pqd_jhscsl":700,"rkd_ttl_rksl_v":600,"rkd_day_rksl_v":480,"err_ttl_gzsl_v":15,"err_day_gzsl_v":15,"bzrs":20}],"Table1":[{"err_gzdm":"AOIG1","err_gzmc":"AOI工站","err_gzsl":20},{"err_gzdm":"DCT","err_gzmc":"DCT工站","err_gzsl":100},{"err_gzdm":"FCT","err_gzmc":"FCT工站","err_gzsl":0},{"err_gzdm":"ICT","err_gzmc":"ICT工站","err_gzsl":100},{"err_gzdm":"MI","err_gzmc":"MI外观检","err_gzsl":50},{"err_gzdm":"MUTAL","err_gzmc":"自互检","err_gzsl":50}],"Table2":[{"pqd_xbdm":"K1","pqd_jhscsl":1400,"rkd_day_rksl_v":500,"pqd_jhdclv":35.71,"pqd_scxlv":83.33,"pqd_yjscxlv":4.17},{"pqd_xbdm":"K2","pqd_jhscsl":1700,"rkd_day_rksl_v":980,"pqd_jhdclv":57.65,"pqd_scxlv":163.33,"pqd_yjscxlv":8.17}]}
      */
 
     private boolean utStatus;
@@ -81,8 +80,8 @@ public class ProductFragmentBean {
              * itm_pmgg : KFR-35G/V1BB1A-3
              * plm_jhsl : 1000
              * pqd_jhscsl : 1000
-             * rkd_ttl_rksl_v : 0
-             * rkd_day_rksl_v : 0
+             * rkd_ttl_rksl_v : 0.0
+             * rkd_day_rksl_v : 0.0
              * err_ttl_gzsl_v : 0
              * err_day_gzsl_v : 0
              * bzrs : 20
@@ -95,8 +94,8 @@ public class ProductFragmentBean {
             private String itm_pmgg;
             private int plm_jhsl;
             private int pqd_jhscsl;
-            private int rkd_ttl_rksl_v;
-            private int rkd_day_rksl_v;
+            private double rkd_ttl_rksl_v;
+            private double rkd_day_rksl_v;
             private int err_ttl_gzsl_v;
             private int err_day_gzsl_v;
             private int bzrs;
@@ -157,19 +156,19 @@ public class ProductFragmentBean {
                 this.pqd_jhscsl = pqd_jhscsl;
             }
 
-            public int getRkd_ttl_rksl_v() {
+            public double getRkd_ttl_rksl_v() {
                 return rkd_ttl_rksl_v;
             }
 
-            public void setRkd_ttl_rksl_v(int rkd_ttl_rksl_v) {
+            public void setRkd_ttl_rksl_v(double rkd_ttl_rksl_v) {
                 this.rkd_ttl_rksl_v = rkd_ttl_rksl_v;
             }
 
-            public int getRkd_day_rksl_v() {
+            public double getRkd_day_rksl_v() {
                 return rkd_day_rksl_v;
             }
 
-            public void setRkd_day_rksl_v(int rkd_day_rksl_v) {
+            public void setRkd_day_rksl_v(double rkd_day_rksl_v) {
                 this.rkd_day_rksl_v = rkd_day_rksl_v;
             }
 
@@ -238,17 +237,17 @@ public class ProductFragmentBean {
             /**
              * pqd_xbdm : K1
              * pqd_jhscsl : 1400
-             * rkd_day_rksl_v : 500
+             * rkd_day_rksl_v : 500.0
              * pqd_jhdclv : 35.71
-             * pqd_scxlv : 125
-             * pqd_yjscxlv : 6.25
+             * pqd_scxlv : 83.33
+             * pqd_yjscxlv : 4.17
              */
 
             private String pqd_xbdm;
             private int pqd_jhscsl;
-            private int rkd_day_rksl_v;
+            private double rkd_day_rksl_v;
             private double pqd_jhdclv;
-            private int pqd_scxlv;
+            private double pqd_scxlv;
             private double pqd_yjscxlv;
 
             public String getPqd_xbdm() {
@@ -267,11 +266,11 @@ public class ProductFragmentBean {
                 this.pqd_jhscsl = pqd_jhscsl;
             }
 
-            public int getRkd_day_rksl_v() {
+            public double getRkd_day_rksl_v() {
                 return rkd_day_rksl_v;
             }
 
-            public void setRkd_day_rksl_v(int rkd_day_rksl_v) {
+            public void setRkd_day_rksl_v(double rkd_day_rksl_v) {
                 this.rkd_day_rksl_v = rkd_day_rksl_v;
             }
 
@@ -283,11 +282,11 @@ public class ProductFragmentBean {
                 this.pqd_jhdclv = pqd_jhdclv;
             }
 
-            public int getPqd_scxlv() {
+            public double getPqd_scxlv() {
                 return pqd_scxlv;
             }
 
-            public void setPqd_scxlv(int pqd_scxlv) {
+            public void setPqd_scxlv(double pqd_scxlv) {
                 this.pqd_scxlv = pqd_scxlv;
             }
